Refactor: GroupRankingViewModel directory 변경
