+++ app/src/main/java/com/asap/aljyo/di/VIewModelFactory.kt
@@ -1,7 +1,7 @@
 package com.asap.aljyo.di
 
 import com.asap.aljyo.core.components.group_details.GroupDetailsViewModel
-import com.asap.aljyo.core.components.group_ranking.GroupRankingViewModel
+import com.asap.aljyo.core.components.viewmodel.GroupRankingViewModel
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityComponent
